feat(modules): phase 2a UI dans le module + squelette 2b, v1.9.1a
Déplace SIA/VAC dans sia-france/renderer, registre IPC renderer bundle,
version 1.9.1a et workflow release avec version personnalisée.
